+++ src/prelude.rs
@@ -18,3 +18,11 @@ pub use structopt::clap::crate_version;
 pub type Result<T, E = Box<dyn Error>> = std::result::Result<T, E>;
 pub type Receiver = crossbeam::channel::Receiver<Message>;
 pub type Sender = crossbeam::channel::Sender<Message>;
+
+/// Amount of [Joule](https://en.wikipedia.org/wiki/Joule)s
+/// in 1 [watt-hour](https://en.wikipedia.org/wiki/Kilowatt-hour).
+pub const JOULES_IN_WH: f64 = 3600.0;
+
+pub const JOULES_IN_KWH: f64 = 1000.0 * JOULES_IN_WH;
+
+pub const WH_IN_JOULE: f64 = 1.0 / JOULES_IN_WH;